refactor: modularize version service and remove frontend build-time version
- Move version service to dedicated src/version module
- Add VERSION_FILE_PATH constant to avoid magic paths
- Extract shared version loading logic to utils
- Remove frontend build-time version.json dependency
- Frontend now fetches version exclusively from API
- Add error handling in UI footer for version unavailability
- Address PR review feedback on code duplication

This change prepares for future frontend/backend separation.
