+++ src/app.module.ts
@@ -3,7 +3,6 @@ import { ConfigModule } from "@nestjs/config";
 import { ServeStaticModule } from "@nestjs/serve-static";
 import { join } from "path";
 import { AppController } from "./app.controller.js";
-import { VersionService } from "./common/version.service.js";
 import { configValidationSchema, loadConfig } from "./config/app.config.js";
 import { DatabaseModule } from "./database/database.module.js";
 import { DataSourceModule } from "./dataSource/dataSource.module.js";
@@ -11,6 +10,7 @@ import { DealModule } from "./deal/deal.module.js";
 import { MetricsModule } from "./metrics/metrics.module.js";
 import { RetrievalModule } from "./retrieval/retrieval.module.js";
 import { SchedulerModule } from "./scheduler/scheduler.module.js";
+import { VersionModule } from "./version/index.js";
 
 @Module({
   imports: [
@@ -29,8 +29,8 @@ import { SchedulerModule } from "./scheduler/scheduler.module.js";
     RetrievalModule,
     DataSourceModule,
     MetricsModule,
+    VersionModule,
   ],
   controllers: [AppController],
-  providers: [VersionService],
 })
 export class AppModule {}
