+++ src/version/index.ts
@@ -1,0 +1,5 @@
+export { VersionModule } from "./version.module.js";
+export { VersionService } from "./version.service.js";
+export { loadVersionInfo } from "./version.utils.js";
+export { VERSION_FILE_PATH } from "./version.constants.js";
+export type { IVersionInfo } from "./version.types.js";
