+++ src/version/version.constants.ts
@@ -1,0 +1,7 @@
+import { join } from "node:path";
+
+/**
+ * Version file path constant
+ * Single source of truth for version.json location
+ */
+export const VERSION_FILE_PATH = join(process.cwd(), "dist", "version.json");
