+++ src/version/version.module.ts
@@ -1,0 +1,8 @@
+import { Module } from "@nestjs/common";
+import { VersionService } from "./version.service.js";
+
+@Module({
+  providers: [VersionService],
+  exports: [VersionService],
+})
+export class VersionModule {}
