+++ src/version/version.types.ts
@@ -1,0 +1,7 @@
+export interface IVersionInfo {
+  version: string;
+  commit: string;
+  commitShort: string;
+  branch: string;
+  buildTime: string;
+}
